requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyascii==0.3.1
apparmor==3.1.2
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autocommand==2.2.2
autopep8==2.0.1
Babel==2.9.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
btrfsutil==6.1.2
CacheControl==0.12.11
cachetools==5.2.1
cachy==0.3.0
caffeine-ng==4.0.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
cleo==2.0.1
click==8.1.3
configobj==5.1.0.dev0
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.4.1
cryptography==38.0.4
cssselect2==0.7.0
cssutils==2.6.0
Cython==0.29.33
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
distro==1.8.0
Django==4.1.2
djoser @ git+https://github.com/sunscrapers/djoser.git@abdf622f95dfa2c6278c4bd6d50dfe69559d90c0
django-allauth==0.52.0
django-braces==1.15.0
django-cors-headers==3.13.0
django-filter==22.1
django-oauth-toolkit==2.2.0
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==1.2.1
drf-yasg==1.21.5
dulwich==0.21.2
duplicity==1.2.1
ecdsa==0.18.0
ewmh==0.1.6
executing==1.2.0
fasteners==0.18
filelock==3.8.2
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
future==0.18.2
GNOME-AppFolders-Manager==0.4.7
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
greenlet==2.0.0
gufw==22.4.0
gunicorn==20.1.0
heroku==0.1.4
html5lib==1.1
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
inflect==6.0.2
inflection==0.5.1
ipython==8.8.0
itsdangerous==2.1.2
itypes==1.2.0
jaraco.classes==3.2.3
jaraco.context==4.2.0
jaraco.functools==3.5.2
jaraco.text==3.11.0
jarowinkler==1.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==1.4.2
keyring==23.11.0
lark==1.1.5
layoutswitcherlib==0.8.33
LibAppArmor==3.1.2
libfdt==1.6.1
lit==15.0.7.dev0
lockfile==0.12.2
lxml==4.9.1
magic-filter==1.0.9
manjaro-sdk==0.1
Markdown==3.4.1
MarkupSafe==2.1.1
material-color-utilities-python==0.1.5
matplotlib-inline==0.1.6
meson==1.0.0
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.4
netsnmp-python==1.0a1
nftables==0.1
npyscreen==4.10.5
nspektr==0.4.0
numpy==1.23.4
oauth==1.0.1
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==21.3
pacman-mirrors==4.23.2
paramiko==2.11.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==2.5.4
ply==3.11
poetry==1.3.2
poetry-core==1.5.0
poetry-plugin-export==1.2.0
prompt-toolkit==3.0.36
protobuf==4.21.12
ptyprocess==0.7.0
pulsectl==22.3.2
pure-eval==0.2.2
pwquality==1.4.5
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.23.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
PyDrive2==1.15.0
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==2.6.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyproject-toml==0.0.10
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.19.3
PySide6==6.4.2
pysmbc==1.0.23
pyTelegramBotAPI==4.10.0
pytesseract==0.3.10
python-dateutil==1.5
python-jose==3.3.0
python-xlib==0.31
python3-openid==3.2.0
pytz==2021.3
pyxdg==0.28
PyYAML==6.0
ranger-fm==1.9.3
rapidfuzz==2.13.7
regex==2022.10.31
reportlab==3.6.12
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.3
setproctitle==1.3.2
shellingham==1.5.0.post1
shiboken6==6.4.2
shiboken6-generator==6.4.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
sqlparse==0.4.3
stack-data==0.6.2
svglib==1.5.1
systemd-python==235
team==1.0
telebot==0.0.4
tinycss2==1.2.1
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
traitlets==5.8.1
trove-classifiers==2023.1.12
typing_extensions==4.4.0
ufw==0.36.1
uritemplate==4.1.1
urllib3==1.26.12
validate==5.1.0.dev0
validate-pyproject==0.10.1
virtualenv==20.17.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==6.3.0
wrapt==1.14.1
Yapsy==1.12.2
yarl==1.8.2
zipp==3.13.0
